controller-api/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
jinja2==3.1.2
kubernetes==28.1.0
kafka-python==2.0.2
python-multipart==0.0.6
pydantic==2.5.0
